fix(angular): add custom wrapper for ion-icon
