+++ angular/src/directives/proxies-list.txt
@@ -29,7 +29,6 @@ export const DIRECTIVES = [
   d.IonFooter,
   d.IonGrid,
   d.IonHeader,
-  d.IonIcon,
   d.IonImg,
   d.IonInfiniteScroll,
   d.IonInfiniteScrollContent,
