+++ angular/src/index.ts
@@ -18,6 +18,7 @@ export { VirtualHeader } from './directives/virtual-scroll/virtual-header';
 export { VirtualFooter } from './directives/virtual-scroll/virtual-footer';
 export { IonModal } from './directives/overlays/modal';
 export { IonPopover } from './directives/overlays/popover';
+export { IonIcon } from './directives/ion-icon';
 export * from './directives/proxies';
 
 // PROVIDERS
